pyproject.toml
[project]
name = "learn-fast-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==3.2.2",
    "fastapi[standard]>=0.120.4",
    "passlib>=1.7.4",
    "pillow>=12.0.0",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "qrcode[pil]>=8.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]
